feat(config-json): Only either bundle or load from remote
