+++ src/globalState.ts
@@ -5,6 +5,7 @@ import type { WorldWarp } from 'flying-squid/dist/lib/modules/warps'
 import type { OptionsGroupType } from './optionsGuiScheme'
 import { appQueryParams } from './appParams'
 import { options, disabledSettings } from './optionsStorage'
+import { AppConfig } from './appConfig'
 
 // todo: refactor structure with support of hideNext=false
 
@@ -110,26 +111,6 @@ export const showContextmenu = (items: ContextMenuItem[], { clientX, clientY }) 
 
 // ---
 
-export type AppConfig = {
-  // defaultHost?: string
-  // defaultHostSave?: string
-  defaultProxy?: string
-  // defaultProxySave?: string
-  // defaultVersion?: string
-  peerJsServer?: string
-  peerJsServerFallback?: string
-  promoteServers?: Array<{ ip, description, version? }>
-  mapsProvider?: string
-
-  appParams?: Record<string, any> // query string params
-
-  defaultSettings?: Record<string, any>
-  forceSettings?: Record<string, boolean>
-  // hideSettings?: Record<string, boolean>
-  allowAutoConnect?: boolean
-  pauseLinks?: Array<Array<Record<string, any>>>
-}
-
 export const miscUiState = proxy({
   currentDisplayQr: null as string | null,
   currentTouch: null as boolean | null,
@@ -144,31 +125,13 @@ export const miscUiState = proxy({
   loadedServerIndex: '',
   /** currently trying to load or loaded mc version, after all data is loaded */
   loadedDataVersion: null as string | null,
-  appLoaded: false,
+  fsReady: false,
   singleplayerAvailable: false,
   usingGamepadInput: false,
   appConfig: null as AppConfig | null,
   displaySearchInput: false,
   displayFullmap: false
 })
-
-export const loadAppConfig = (appConfig: AppConfig) => {
-  if (miscUiState.appConfig) {
-    Object.assign(miscUiState.appConfig, appConfig)
-  } else {
-    miscUiState.appConfig = appConfig
-  }
-
-  if (appConfig.forceSettings) {
-    for (const [key, value] of Object.entries(appConfig.forceSettings)) {
-      if (value) {
-        disabledSettings.value.delete(key)
-      } else {
-        disabledSettings.value.add(key)
-      }
-    }
-  }
-}
 
 export const isGameActive = (foregroundCheck: boolean) => {
   if (foregroundCheck && activeModalStack.length) return false
